Finalizado implementação novo menu #7
